+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ##2020.08.1
 * added `CPUTime` and `AffectedShards` fields to `QueryPhase` struct
-* added `CompilationPhase` statistics
+* added `CompilationStats` statistics
 
 ##2020.07.7 
 * support manage table attributes